+++ src/ts/ui/ui.ts
@@ -16,17 +16,17 @@ export const setupButton = (element: HTMLImageElement, path: string, onclick: ()
 
 	element.src = normal;
 	element.addEventListener('mouseenter', () => {
+		hovered = true;
+		element.setAttribute('data-hovered', '');
 		if (element.style.pointerEvents === 'none') return;
 		if (!element.hasAttribute('data-locked')) element.src = held? down : hover;
 		if (!held) AudioManager.play('buttonover.wav');
-		hovered = true;
-		element.setAttribute('data-hovered', '');
 	});
 	element.addEventListener('mouseleave', () => {
-		if (element.style.pointerEvents === 'none') return;
-		if (!element.hasAttribute('data-locked')) element.src = normal;
 		hovered = false;
 		element.removeAttribute('data-hovered');
+		if (element.style.pointerEvents === 'none') return;
+		if (!element.hasAttribute('data-locked')) element.src = normal;		
 	});
 	element.addEventListener('mousedown', (e) => {
 		if (element.style.pointerEvents === 'none') return;
@@ -37,8 +37,8 @@ export const setupButton = (element: HTMLImageElement, path: string, onclick: ()
 		if (triggerOnMouseDown) onclick();
 	});
 	window.addEventListener('mouseup', () => {
-		if (element.style.pointerEvents === 'none') return;
 		held = false;
+		if (element.style.pointerEvents === 'none') return;
 		if (!element.hasAttribute('data-locked')) element.src = hovered? hover : normal;
 	});
 	if (!triggerOnMouseDown) element.addEventListener('click', (e) => e.button === 0 && onclick());